Reworked how database connection service works.
